chore: folder structure refactor
