+++ packages/@lwc/wds-core/src/node/plugins/mock-esm/handle/controller.js
@@ -1,8 +1,8 @@
 import { relative as pathRelative, resolve as pathResolve } from 'node:path';
 import { stringify as qsStringify } from 'node:querystring';
 import resolveSync from 'resolve/sync.js';
-import { toolkitSrcPath } from '../../util.js';
-import { MOCK_CONTROLLER_PREFIX, MOCK_STUB_PREFIX } from './const.js';
+import { toolkitSrcPath } from '../../../util.js';
+import { MOCK_CONTROLLER_PREFIX, MOCK_STUB_PREFIX } from '../const.js';
 
 const LEXER_ABS_PATH = resolveSync('es-module-lexer/dist/lexer.asm.js', {
   baseDir: toolkitSrcPath,
